web: request status frame
